requirements.txt
boto3
requests
AWSIoTPythonSDK
awesome-slugify
